Consolate classes
