+++ spec/stack_master/commands/apply_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe StackMaster::Commands::Apply do
     allow(Aws::CloudFormation::Client).to receive(:new).and_return(cf)
     allow(cf).to receive(:update_stack)
     allow(cf).to receive(:create_stack)
-    allow(StackMaster::StackDiffer).to receive(:perform).with(proposed_stack, stack)
+    allow(StackMaster::StackDiffer).to receive(:new).with(proposed_stack, stack).and_return double.as_null_object
     allow(STDOUT).to receive(:print)
     allow(STDIN).to receive(:getch).and_return('y')
     allow(StackMaster::StackEvents::Streamer).to receive(:stream)
