+++ spec/stack_master/stack_differ_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe StackMaster::StackDiffer do
-  subject(:stack_master) { described_class.new(proposed_stack, stack) }
+  subject(:differ) { described_class.new(proposed_stack, stack) }
   let(:stack) { StackMaster::Stack.new(stack_name: stack_name,
                                        region: region,
                                        stack_id: 123,
@@ -12,22 +12,22 @@ RSpec.describe StackMaster::StackDiffer do
   let(:stack_name) { 'myapp-vpc' }
   let(:region) { 'us-east-1' }
 
-  describe "#perform" do
+  describe "#output_diff" do
     context "entirely new stack" do
       let(:stack) { nil }
 
       it "outputs the entire stack" do
-        expect { stack_master.perform }.to output(/\+  \"a\"\: 1/).to_stdout
-        expect { stack_master.perform }.to output(/\+  \"param1\"\: \"hello\"/).to_stdout
-        expect { stack_master.perform }.to output(/No stack found/).to_stdout
+        expect { differ.output_diff }.to output(/\+  \"a\"\: 1/).to_stdout
+        expect { differ.output_diff }.to output(/\+  \"param1\"\: \"hello\"/).to_stdout
+        expect { differ.output_diff }.to output(/No stack found/).to_stdout
       end
     end
 
     context "stack update" do
       it "outputs the stack diff" do
-        expect { stack_master.perform }.to output(/\+  \"a\"\: 1/).to_stdout
-        expect { stack_master.perform }.to output(/\+  \"param1\"\: \"hello\"/).to_stdout
-        expect { stack_master.perform }.to_not output(/No stack found/).to_stdout
+        expect { differ.output_diff }.to output(/\+  \"a\"\: 1/).to_stdout
+        expect { differ.output_diff }.to output(/\+  \"param1\"\: \"hello\"/).to_stdout
+        expect { differ.output_diff }.to_not output(/No stack found/).to_stdout
       end
     end
   end
